Split `File` -> `FileWithDependency`
