+++ src/Files/HasDependency.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace BrianHenryIE\Strauss\Files;
+
+use BrianHenryIE\Strauss\Composer\ComposerPackage;
+
+interface HasDependency
+{
+
+    public function getDependency(): ComposerPackage;
+
+    public function getTargetRelativePath(): string;
+
+    /**
+     * Record the autoloader it is found in. Which could be all of them.
+     */
+    public function addAutoloader(string $autoloaderType): void;
+
+    public function isFilesAutoloaderFile(): bool;
+}
